+++ CHANGELOG.md
@@ -9,6 +9,15 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+### Fixed
+
+- Fix printing of single line descriptions with backslashes
+
+### Changed
+
+- Print long argument lists on multiple lines
+- Print space between object value brackets
+
 ## v15.0.0
 
 ### Changed